fix: three defects found by the local end-to-end RAG test
Booted the app locally (port 8099) and exercised the full path against
the dedicated agent_assist_docs_openai collection on the 74 Qdrant:
upload -> list -> /process (grounded suggestion) -> pause -> re-ask ->
resume -> download -> delete. Three real defects surfaced and fixed:

1. Ingestion failed outright: M4's QdrantValueFactory rejects Long
   metadata ("Unsupported Qdrant value type"). uploadTimestamp is now
   stored as a Double - Qdrant range filters compare numerics, so date
   filtering semantics are unchanged.

2. EVERY first message of a new conversation 500'd under MySQL's
   default REPEATABLE READ: the outer transaction's snapshot cannot see
   the conversation row the REQUIRES_NEW ConversationCreator just
   committed, so the second getOrCreate re-inserted, hit the duplicate
   key, and the recovery refetch was equally blind ("Conversation not
   found after race condition"). processMessage now runs READ_COMMITTED.
   Pre-existing bug from 6d93170 - deployments on READ-COMMITTED MySQL
   never noticed.

3. LLM context and source previews contained the metadata map instead
   of document text (M4's getFormattedContent prepends all metadata,
   unlike upstream's version) - switched to getContent().

Verified end-to-end: suggestion grounded in the uploaded FAQ (opening
time, price, dress code all correct), paused document excluded with the
registry-served no-knowledge reply, resume instant, download serves the
original file, delete cleans Qdrant + storage. Suite: 92 green.

+++ src/main/java/com/agentassist/rag/AgentAssistDocumentService.java
@@ -172,7 +172,10 @@ public class AgentAssistDocumentService {
             d.getMetadata().put("source", getFileExtension(fileName));
             d.getMetadata().put("companyId", String.valueOf(companyId));
             d.getMetadata().put("uploadDate", uploadDate);
-            d.getMetadata().put("uploadTimestamp", uploadTimestamp);
+            // Double, not long: M4's QdrantValueFactory rejects Long metadata
+            // ("Unsupported Qdrant value type") — found by the local smoke test.
+            // Qdrant range filters compare numerics, so date filtering still works.
+            d.getMetadata().put("uploadTimestamp", (double) uploadTimestamp);
             d.getMetadata().put("status", "UPLOADED");
             d.getMetadata().put("useCase", USE_CASE_AGENT_ASSIST);  // Tag for agent assist
             // Searchable by default. Pausing sets this to "false"; retrieval excludes
